Move map files + Remove hover effect + add sample image to markers
